+++ assets/javascripts/kitten/components/molecules/alert-box/test.js
@@ -1,0 +1,10 @@
+import React from 'react'
+import renderer from 'react-test-renderer'
+import { AlertBox } from './'
+
+describe('<AlertBox />', () => {
+  it('should matches with default snapshot', () => {
+    const component = renderer.create(<AlertBox>Alert !</AlertBox>).toJSON()
+    expect(component).toMatchSnapshot()
+  })
+})